fixed testUpdateBeerById
Needed to flush manually the Beer entity updated before executing the findById. This is needed due to the version not updated correctly. The version is updated when the entity is flushed to the db. The entity is flushed at the end of a transaction (in this case when the test is finished) or before JPQL, HQL or SQL queries.