Updated entity annotation for uuid
Initially hibernate behaviour was overwritten by the @Column annotation -> field: columnDefinition="varchar". Mysql needs a lenght. so "varchar" changed to "varchar(36). After that the problem was that hibernate was storing the uuid as a binary and not a string. so added the annotation @JdbcTypeCode(SqlTypes.CHAR) on the id property in Beer and Customer entities.